Reduce Jitter
Previous code produced JSON either too fast, or too slow, can't really
tell. Either way it was stuttering the bars. Now the ESP will send a
"passcode" for the server to acknowldge it is the only source of data,
and does not send it back. Perhaps saving ESP cpu clocks. And making
graphs MUCH smoother.